requirements.txt
# Global AI Hackathon - Common Requirements
# Comprehensive package list for rapid AI/ML prototyping

# === Core Data Science ===
numpy>=1.24
pandas>=2.1
scikit-learn>=1.3
scipy>=1.10

# === LLM & AI APIs ===
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-mcp-adapters>=0.0.3
mcp>=1.1.0
langchain-ollama>=0.1.0

# === Agentic Frameworks ===
langgraph>=0.0.20
# autogen-agentchat>=0.2.0  # Uncomment for Microsoft AutoGen

# === Hugging Face Ecosystem ===
huggingface-hub>=0.20.0
# transformers[agents]>=4.37.0  # For HF Agents (installs extras)

# === Deep Learning & Transformers ===
torch>=2.2.0
transformers>=4.37.0
accelerate>=0.26.0
sentence-transformers>=2.3.0

# === Vector Databases & Embeddings ===
chromadb>=0.4.22
faiss-cpu>=1.7.4
# pinecone-client>=3.0.0  # Uncomment if using Pinecone

# === Web UI & APIs ===
streamlit>=1.31.0
gradio>=4.16.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# === Computer Vision ===
opencv-python>=4.9.0
Pillow>=10.2.0

# === Audio Processing ===
elevenlabs>=1.0.0
# openai-whisper>=20231117  # Uncomment for speech-to-text

# === Utilities ===
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.6.0
pydot>=2.0.0
pygraphviz>=1.11

# === Dev Tools (optional) ===
# black>=24.1.0
# ruff>=0.2.0
# pytest>=8.0.0

# Note: For GPU support, install PyTorch from https://pytorch.org
# For production, pin exact versions: pip freeze > requirements.txt
